pyproject.toml
[tool.poetry]
name = "sith-be"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-stubs = "1.9.0"
Django = "3.2.10"
graphene-stubs = "0.15"
graphene-django = "2.15.0"
invoke = "1.6.0"
mypy-extensions = "0.4.3"
psycopg2 = "2.9.2"
sqlparse = "0.4.2"
toml = "0.10.2"
types-pytz = "2021.3.0"
types-PyYAML = "6.0.1"
typing-extensions = "4.0.1"
django-cors-headers = "3.10.1"
pytest-django = "^4.5.2"
snapshottest = "^0.6.0"

[tool.poetry.dev-dependencies]
lxml = "^4.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "server.settings"
minversion = "6.0"
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py",
]
